Tests for reward contracts moved to a dedicated directory
